+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-- Better handling of aetheryte shortcuts -alydev
-- Stop After Quest button -Kage
+- Bug fix: don't teleport if Transporting status -alydev
+- Aetheryte shortcut: wait until not occupied -alydev
+- Stop After Quest, Stop After Teleport button -Kage
+  - button changed to Stop Before Teleport -Kage
+- Buttons organised slightly -alydev
+- Paths checked -alydev